# storage/api-docs/api-docs.json
{"swagger":"2.0","info":{"title":"laravel\u9879\u76ee\u63a5\u53e3\u6587\u6863","description":"laravel\u9879\u76eeAPI\u63a5\u53e3\u6587\u6863","contact":{"name":"laravelProject","email":"rachel.ran@jfz.com"},"version":"1.0.0"},"host":"www.laravelproject.net","schemes":["http","https"],"paths":{"\/component\/detail":{"get":{"tags":["component\/detail"],"description":"\u83b7\u53d6component\u7684\u8be6\u7ec6\u4fe1\u606f","operationId":"component.detail","produces":["application\/json"],"parameters":[{"name":"fundDcId","in":"path","description":"\u91d1\u878d\u6570\u636e\u4e2d\u5fc3\u552f\u4e00id","type":"string"}],"responses":{"200":{"description":"\u6210\u529f\u54cd\u5e94","schema":{"properties":{"code":{"description":"\u4e1a\u52a1\u903b\u8f91\u72b6\u6001\u7801","type":"integer"},"msg":{"description":"\u4e1a\u52a1\u903b\u8f91\u5904\u7406\u7ed3\u679c\u4fe1\u606f","type":"string"},"content":{"properties":{"data":{"description":"\u54cd\u5e94\u6570\u636e","properties":{"id":{"description":"id","type":"integer"},"fundDcId":{"description":"\u91d1\u878d\u6570\u636e\u4e2d\u5fc3\u552f\u4e00id","type":"string"},"name":{"description":"\u4fdd\u9669\u4ea7\u54c1\u7b80\u79f0","type":"string"},"fname":{"description":"\u4fdd\u9669\u4ea7\u54c1\u5168\u79f0","type":"string"}}}}}},"type":"json"}}}}},"\/demo\/mail":{"get":{"tags":["mail"],"summary":"\u5355\u7528\u6237\u8be6\u7ec6\u4fe1\u606f","description":"\u8fd4\u56de\u5355\u4e2a\u7528\u6237\u7684\u8be6\u7ec6\u4fe1\u606f","operationId":"demo.mail","produces":["application\/json"],"responses":{"200":{"description":"\u8fd4\u56de\u56de\u6210\u529f\u76f8\u5173\u8bf4\u660e"},"401":{"description":"\u8fd4\u56de\u5931\u8d25\u76f8\u5173\u8bf4\u660e"}}}},"\/demo\/member\/get":{"get":{"tags":["member"],"summary":"\u5355\u7528\u6237\u8be6\u7ec6\u4fe1\u606f","description":"\u8fd4\u56de\u5355\u4e2a\u7528\u6237\u7684\u8be6\u7ec6\u4fe1\u606f","operationId":"demo.member.get","produces":["application\/json"],"responses":{"200":{"description":"\u8fd4\u56de\u56de\u6210\u529f\u76f8\u5173\u8bf4\u660e"},"401":{"description":"\u8fd4\u56de\u5931\u8d25\u76f8\u5173\u8bf4\u660e"}}}},"\/demo\/userdetail\/{id}":{"get":{"tags":["users"],"summary":"\u5355\u7528\u6237\u8be6\u7ec6\u4fe1\u606f","description":"\u8fd4\u56de\u5355\u4e2a\u7528\u6237\u7684\u8be6\u7ec6\u4fe1\u606f","operationId":"demo.userdetail","produces":["application\/json"],"parameters":[{"name":"","in":"query","description":"\u63a5\u53e3\u9700\u8981\u7684\u53c2\u76f8\u5173\u4fe1\u606f\uff0c\u5b57\u6bb5\u540d\uff0c\u4f20\u53c2\u65b9\u5f0f\uff0c\u662f\u5426\u5fc5\u987b\uff0c\u7c7b\u578b\u7b49","required":true,"type":"integer"}],"responses":{"200":{"description":"\u8fd4\u56de\u56de\u6210\u529f\u76f8\u5173\u8bf4\u660e"},"401":{"description":"\u8fd4\u56de\u5931\u8d25\u76f8\u5173\u8bf4\u660e"}}}},"\/demo\/index":{"get":{"tags":["users"],"summary":"\u7528\u6237\u8d44\u6599\u7ba1\u7406","description":"\u7528\u6237\u8d44\u6599\u5217\u8868","operationId":"user.index","produces":["application\/json"],"parameters":[{"name":"page","in":"query","description":"\u5206\u9875\u7f16\u53f7,\u9ed8\u8ba41","type":"integer"},{"name":"pagesize","in":"query","description":"\u6bcf\u9875\u663e\u793a\u6761\u6570,\u9ed8\u8ba410","type":"integer"}],"responses":{"200":{"description":"\u7528\u6237\u5217\u8868","schema":{"properties":{"pages":{"properties":{"totalnum":{"description":"\u8bb0\u5f55\u603b\u6570","type":"integer"},"totalpage":{"description":"\u603b\u9875\u6570","type":"integer"},"pagesize":{"description":"\u6bcf\u9875\u663e\u793a\u8bb0\u5f55\u6570","type":"integer"},"page":{"description":"\u5f53\u524d\u9875\u6570","type":"integer"}}},"lists":{"items":{"properties":{"id":{"description":"id","type":"integer"},"source":{"description":"\u6570\u636e\u6765\u6e90","type":"string"},"industry_no":{"description":"\u884c\u4e1a\u7f16\u53f7","type":"integer"},"channel_id":{"description":"\u6e20\u9053id","type":"integer"},"name":{"description":"\u540d\u79f0","type":"string"}}}}},"type":"json"}},"422":{"description":"error"}}}},"\/demo\/list":{"get":{"tags":["users"],"summary":"\u6d4b\u8bd5\u7f16\u8f91\u5668\u751f\u6210\u6a21\u677f","description":"\u6d4b\u8bd5\u7f16\u8f91\u5668\u751f\u6210\u6a21\u677f","operationId":"demo.list","produces":["application\/json"],"parameters":[{"name":"id","in":"query","description":"\u7528\u6237id","required":true,"type":"integer"}],"responses":{"200":{"description":"\u6d4b\u8bd5\u6210\u529f","schema":{"properties":{"list":{"properties":{"name":{"description":"\u7528\u6237\u59d3\u540d","type":"string"}}}},"type":"json"}}}}},"\/demo\/show":{"get":{"tags":["users"],"summary":"\u6d4b\u8bd5\u7f16\u8f91\u5668\u6a21\u677f","description":"\u6d4b\u8bd5\u7f16\u8f91\u5668\u6a21\u677f","operationId":"demo.show","produces":["application\/json"],"parameters":[{"name":"id","in":"query","description":"id","required":true,"type":"integer"},{"name":"name","in":"query","description":"\u59d3\u540d","required":true,"type":"string"}],"responses":{"200":{"description":"success","schema":{"properties":{"list1":{"properties":{"id":{"description":"\u7528\u6237id","type":"integer"},"name":{"description":"\u7528\u6237\u59d3\u540d","type":"string"}}},"list2":{"properties":{"age":{"description":"\u5e74\u9f84","type":"integer"},"sex":{"description":"\u6027\u522b1\u75372\u5973","type":"integer"}}}},"type":"json"}}}}},"\/company\/get":{"get":{"tags":["users"],"summary":"\u6240\u6709\u7528\u6237\u76f8\u5173\u4fe1\u606f","description":"\u8fd4\u56de\u6240\u6709\u7528\u6237\u7684\u76f8\u5173\u57fa\u672c\u4fe1\u606f","operationId":"demo.users","produces":["application\/json"],"parameters":[{"name":"tags","in":"query","description":"\u63a5\u53e3\u9700\u8981\u7684\u53c2\u76f8\u5173\u4fe1\u606f\uff0c\u5b57\u6bb5\u540d\uff0c\u4f20\u53c2\u65b9\u5f0f\uff0c\u662f\u5426\u5fc5\u987b\uff0c\u7c7b\u578b\u7b49","required":true,"type":"string"}],"responses":{"200":{"description":"\u8fd4\u56de\u56de\u6210\u529f\u76f8\u5173\u8bf4\u660e"},"401":{"description":"\u8fd4\u56de\u5931\u8d25\u76f8\u5173\u8bf4\u660e"}}}}},"definitions":{},"securityDefinitions":{"client":{"type":"apiKey","description":"client credentials grant type","name":"Authorization","in":"header"}},"security":[{"client":[]}]}